feat: move utils to base embedder
